requirements.txt
boto3==1.14.1
python-dotenv==0.13.0
